Added multi-string support. Revealed bug. Read description!
The strings are grabbed twice for some reason. (happened before my commits too)

This wasn't  a problem before because the strings remained the same. But now it's obvious for the user that something is wrong because the first set of strings only flashes for a split second and then the second set of strings appears. (Not always though. It depends on how fast the phone is.)

Check out the commented lines (215 and 216). Those will clearly show the thing happening.